Extracted get conversation
