+++ services/galley/test/integration/API.hs
@@ -110,23 +110,15 @@ tests s =
         [ test s "status" status,
           test s "metrics" metrics,
           test s "fetch conversation by qualified ID (v2)" testGetConvQualifiedV2,
-          test s "get empty conversations" getConvsOk,
-          test s "get conversations by ids" getConvsOk2,
-          test s "fail to get >500 conversations with v2 API" getConvsFailMaxSizeV2,
-          test s "get conversation ids with v2 API" testGetConvIdsV2,
           test s "list conversation ids" testListConvIds,
           test s "paginate through conversation ids with v2 API" paginateConvIds,
           test s "paginate through /conversations/list-ids" paginateConvListIds,
           test s "paginate through /conversations/list-ids - page ending at locals and remote domain" paginateConvListIdsPageEndingAtLocalsAndDomain,
-          test s "fail to get >1000 conversation ids" getConvIdsFailMaxSize,
-          test s "fail to get >1000 conversation ids with v2 API" getConvIdsFailMaxSizeV2,
           test s "page through conversations" getConvsPagingOk,
           test s "upgrade connect/invite conversation" putConvAcceptOk,
           test s "upgrade conversation retries" putConvAcceptRetry,
           test s "repeat / cancel connect requests" postRepeatConnectConvCancel,
           test s "block/unblock a connect/1-1 conversation" putBlockConvOk,
-          test s "get conversation" getConvOk,
-          test s "get qualified conversation" getConvQualifiedOk,
           test s "conversation meta access" accessConvMeta,
           test s "add members" postMembersOk,
           test s "add existing members" postMembersOk2,
@@ -136,12 +128,6 @@ tests s =
           test s "fail to add members when not connected" postMembersFail,
           test s "fail to add too many members" postTooManyMembersFail,
           test s "add remote members" testAddRemoteMember,
-          test s "get conversations/:domain/:cnv - local" testGetQualifiedLocalConv,
-          test s "get conversations/:domain/:cnv - local, not found" testGetQualifiedLocalConvNotFound,
-          test s "get conversations/:domain/:cnv - local, not participating" testGetQualifiedLocalConvNotParticipating,
-          test s "get conversations/:domain/:cnv - remote" testGetQualifiedRemoteConv,
-          test s "get conversations/:domain/:cnv - remote, not found" testGetQualifiedRemoteConvNotFound,
-          test s "get conversations/:domain/:cnv - remote, not found on remote" testGetQualifiedRemoteConvNotFoundOnRemote,
           test s "post conversations/list/v2" testBulkGetQualifiedConvs,
           test s "add remote members on invalid domain" testAddRemoteMemberInvalidDomain,
           test s "add remote members when federation isn't enabled" testAddRemoteMemberFederationDisabled,
@@ -723,84 +709,6 @@ postJoinConvFail = do
   conv <- decodeConvId <$> postConv alice [] (Just "gossip") [] Nothing Nothing
   void $ postJoinConv bob conv !!! const 403 === statusCode
 
-getConvsOk :: TestM ()
-getConvsOk = do
-  usr <- randomUser
-  convs <- getAllConvs usr
-  liftIO $
-    [selfConv usr, mlsSelfConvId usr]
-      @?= map (qUnqualified . cnvQualifiedId) convs
-
-getConvsOk2 :: TestM ()
-getConvsOk2 = do
-  [alice, bob] <- randomUsers 2
-  connectUsers alice (singleton bob)
-  -- create & get one2one conv
-  cnv1 <- responseJsonError =<< postO2OConv alice bob (Just "gossip1") <!! const 200 === statusCode
-  do
-    r <-
-      responseJsonError
-        =<< getConvs alice [cnvQualifiedId cnv1] <!! do
-          const 200 === statusCode
-    liftIO $
-      [cnvQualifiedId cnv1] @=? map cnvQualifiedId (crFound r)
-  -- create & get group conv
-  carl <- randomUser
-  connectUsers alice (singleton carl)
-  cnv2 <-
-    responseJsonError
-      =<< postConv alice [bob, carl] (Just "gossip2") [] Nothing Nothing
-        <!! const 201 === statusCode
-  do
-    r <-
-      responseJsonError
-        =<< getConvs alice [cnvQualifiedId cnv2] <!! do
-          const 200 === statusCode
-    liftIO $
-      [cnvQualifiedId cnv2] @=? map cnvQualifiedId (crFound r)
-  -- get both
-  convs <- getAllConvs alice
-  let c1 = find ((== cnvQualifiedId cnv1) . cnvQualifiedId) convs
-  let c2 = find ((== cnvQualifiedId cnv2) . cnvQualifiedId) convs
-  liftIO . forM_ [(cnv1, c1), (cnv2, c2)] $ \(expected, actual) -> do
-    assertEqual
-      "name mismatch"
-      (Just $ cnvName expected)
-      (cnvName <$> actual)
-    assertEqual
-      "self member mismatch"
-      (Just . cmSelf $ cnvMembers expected)
-      (cmSelf . cnvMembers <$> actual)
-    assertEqual
-      "other members mismatch"
-      (Just [])
-      ((\c -> cmOthers (cnvMembers c) \\ cmOthers (cnvMembers expected)) <$> actual)
-
-getConvsFailMaxSizeV2 :: TestM ()
-getConvsFailMaxSizeV2 = do
-  usr <- randomUser
-  g <- view tsUnversionedGalley
-  get
-    ( g
-        . path "/v2/conversations"
-        . zUser usr
-        . zConn "conn"
-        . queryItem "size" (toByteString' (501 :: Int32))
-    )
-    !!! const 400 === statusCode
-
-testGetConvIdsV2 :: TestM ()
-testGetConvIdsV2 = do
-  [alice, bob] <- randomUsers 2
-  connectUsers alice (singleton bob)
-  void $ postO2OConv alice bob (Just "gossip")
-  getConvIdsV2 alice Nothing Nothing !!! do
-    const 200 === statusCode
-    const 2 === length . decodeConvIdList
-  getConvIdsV2 bob Nothing Nothing !!! do
-    const 200 === statusCode
-    const 2 === length . decodeConvIdList
-
 paginateConvIds :: TestM ()
 paginateConvIds = do
   [alice, bob, eve] <- randomUsers 3
@@ -828,18 +736,6 @@ paginateConvIds = do
           else assertEqual ("hasMore should be False, " <> show n <> " more chunks to go") False (convHasMore c)
 
       pure (Just (Right (last (convList c))))
-
-getConvIdsFailMaxSizeV2 :: TestM ()
-getConvIdsFailMaxSizeV2 = do
-  usr <- randomUser
-  getConvIdsV2 usr Nothing (Just 1001)
-    !!! const 400 === statusCode
-
-getConvIdsFailMaxSize :: TestM ()
-getConvIdsFailMaxSize = do
-  usr <- randomUser
-  getConvPage usr Nothing (Just 1001)
-    !!! const 400 === statusCode
 
 testListConvIds :: TestM ()
 testListConvIds = do
@@ -1133,35 +1029,6 @@ putBlockConvOk = do
   getConvQualified bob qconvId !!! do
     const 200 === statusCode
 
-getConvOk :: TestM ()
-getConvOk = do
-  alice <- randomUser
-  bob <- randomUser
-  chuck <- randomUser
-  connectUsers alice (list1 bob [chuck])
-  conv <- decodeConvId <$> postConv alice [bob, chuck] (Just "gossip") [] Nothing Nothing
-  getConv alice conv !!! const 200 === statusCode
-  getConv bob conv !!! const 200 === statusCode
-  getConv chuck conv !!! const 200 === statusCode
-
-getConvQualifiedOk :: TestM ()
-getConvQualifiedOk = do
-  alice <- randomUser
-  bob <- randomQualifiedUser
-  chuck <- randomQualifiedUser
-  connectLocalQualifiedUsers alice (list1 bob [chuck])
-  conv <-
-    decodeConvId
-      <$> postConvQualified
-        alice
-        defNewProteusConv
-          { newConvQualifiedUsers = [bob, chuck],
-            newConvName = checked "gossip"
-          }
-  getConv alice conv !!! const 200 === statusCode
-  getConv (qUnqualified bob) conv !!! const 200 === statusCode
-  getConv (qUnqualified chuck) conv !!! const 200 === statusCode
-
 accessConvMeta :: TestM ()
 accessConvMeta = do
   g <- viewGalley
@@ -1238,96 +1105,6 @@ testAddRemoteMember = do
           "on-new-remote-conversation" ~> EmptyResponse,
           "on-conversation-updated" ~> ()
         ]
-
-testGetQualifiedLocalConv :: TestM ()
-testGetQualifiedLocalConv = do
-  alice <- randomUser
-  convId <- decodeQualifiedConvId <$> postConv alice [] (Just "gossip") [] Nothing Nothing
-  conv :: Conversation <- fmap responseJsonUnsafe $ getConvQualified alice convId <!! const 200 === statusCode
-  liftIO $ do
-    assertEqual "conversation id" convId (cnvQualifiedId conv)
-    assertEqual "conversation name" (Just "gossip") (cnvName conv)
-
-testGetQualifiedLocalConvNotFound :: TestM ()
-testGetQualifiedLocalConvNotFound = do
-  alice <- randomUser
-  localDomain <- viewFederationDomain
-  convId <- (`Qualified` localDomain) <$> randomId
-  getConvQualified alice convId !!! do
-    const 404 === statusCode
-    const (Right (Just "no-conversation")) === fmap (view (at "label")) . responseJsonEither @Object
-
-testGetQualifiedLocalConvNotParticipating :: TestM ()
-testGetQualifiedLocalConvNotParticipating = do
-  alice <- randomUser
-  bob <- randomUser
-  convId <- decodeQualifiedConvId <$> postConv bob [] (Just "gossip about alice") [] Nothing Nothing
-  getConvQualified alice convId !!! do
-    const 403 === statusCode
-    const (Just "access-denied") === view (at "label") . responseJsonUnsafe @Object
-
-testGetQualifiedRemoteConv :: TestM ()
-testGetQualifiedRemoteConv = do
-  aliceQ <- randomQualifiedUser
-  let aliceId = qUnqualified aliceQ
-  loc <- flip toLocalUnsafe () <$> viewFederationDomain
-  bobId <- randomId
-  convId <- randomId
-  let remoteDomain = Domain "far-away.example.com"
-      bobQ = Qualified bobId remoteDomain
-      remoteConvId = Qualified convId remoteDomain
-      bobAsOtherMember = OtherMember bobQ Nothing roleNameWireAdmin
-      aliceAsLocal =
-        LocalMember aliceId defMemberStatus Nothing roleNameWireAdmin
-      aliceAsOtherMember = localMemberToOther (qDomain aliceQ) aliceAsLocal
-      aliceAsSelfMember = localMemberToSelf loc aliceAsLocal
-
-  connectWithRemoteUser aliceId bobQ
-  registerRemoteConv remoteConvId bobId Nothing (Set.fromList [aliceAsOtherMember])
-
-  let mockConversation = mkProteusConv convId bobId roleNameWireAdmin [bobAsOtherMember]
-      remoteConversationResponse = GetConversationsResponse [mockConversation]
-      expected =
-        Conversation
-          remoteConvId
-          (rcnvMetadata mockConversation)
-          (ConvMembers aliceAsSelfMember (rcmOthers (rcnvMembers mockConversation)))
-          ProtocolProteus
-
-  (respAll, _) <-
-    withTempMockFederator'
-      (mockReply remoteConversationResponse)
-      (getConvQualified aliceId remoteConvId)
-
-  conv <- responseJsonUnsafe <$> (pure respAll <!! const 200 === statusCode)
-  liftIO $ do assertEqual "conversation metadata" expected conv
-
-testGetQualifiedRemoteConvNotFound :: TestM ()
-testGetQualifiedRemoteConvNotFound = do
-  aliceId <- randomUser
-  let remoteDomain = Domain "far-away.example.com"
-  remoteConvId <- (`Qualified` remoteDomain) <$> randomId
-  -- No need to mock federator as we don't expect a call to be made
-  getConvQualified aliceId remoteConvId !!! do
-    const 404 === statusCode
-    const (Just "no-conversation") === view (at "label") . responseJsonUnsafe @Object
-
-testGetQualifiedRemoteConvNotFoundOnRemote :: TestM ()
-testGetQualifiedRemoteConvNotFoundOnRemote = do
-  aliceQ <- randomQualifiedUser
-  let aliceId = qUnqualified aliceQ
-  bobId <- randomId
-  convId <- randomId
-  let remoteDomain = Domain "far-away.example.com"
-      remoteConvId = Qualified convId remoteDomain
-      aliceAsOtherMember = OtherMember aliceQ Nothing roleNameWireAdmin
-
-  registerRemoteConv remoteConvId bobId Nothing (Set.fromList [aliceAsOtherMember])
-
-  void . withTempMockFederator' (mockReply (GetConversationsResponse [])) $ do
-    getConvQualified aliceId remoteConvId !!! do
-      const 404 === statusCode
-      const (Just "no-conversation") === view (at "label") . responseJsonUnsafe @Object
 
 -- | Tests getting many converations given their ids.
 --
